feat: consolidate current views and examples. extract shared parts.
