+++ src/components/eip3643/user/BalanceOfERC20.tsx
@@ -1,0 +1,73 @@
+import { useWalletInterface } from "@/services/wallets/useWalletInterface";
+import { useState } from "react";
+import {
+  Button,
+  Flex,
+  Heading,
+  HStack,
+  Input,
+  Stack,
+  Text,
+  VStack,
+} from "@chakra-ui/react";
+import { AccountId, ContractId } from "@hashgraph/sdk";
+import { ContractFunctionParameterBuilder } from "@/services/wallets/contractFunctionParameterBuilder";
+import { appConfig } from "@/config";
+import { convertAccountIdToSolidityAddress } from "@/services/util/helpers";
+
+export default function BalanceOfERC20() {
+  const { accountId, walletName, walletInterface } = useWalletInterface();
+  const [txId, setTxId] = useState("no transaction initiated");
+  const [fungibleTokenEvmAddress, setFungibleTokenEvmAddress] = useState(
+    "0x0000000000000000000000000000000000387719",
+  );
+
+  return (
+    <VStack gap={2} alignItems="flex-start">
+      <Heading size={"md"}>
+        Balance of fungible TestToken ERC20 CA call
+        0x0000000000000000000000000000000000387719
+      </Heading>
+
+      <VStack alignItems="flex-start">
+        <label htmlFor="erc20-balance">Token ID</label>
+        <Input
+          name="erc20-balance"
+          value={fungibleTokenEvmAddress}
+          onChange={(e) => setFungibleTokenEvmAddress(e.target.value)}
+          sx={{
+            minWidth: "400px",
+          }}
+        />
+      </VStack>
+
+      <Button
+        onClick={async () => {
+          setTxId("waiting...");
+
+          const txId = await walletInterface?.executeContractFunction(
+            ContractId.fromEvmAddress(0, 0, fungibleTokenEvmAddress),
+            "balanceOf",
+            new ContractFunctionParameterBuilder().addParam({
+              type: "address",
+              name: "",
+              value: convertAccountIdToSolidityAddress(
+                AccountId.fromString(accountId),
+              ),
+            }),
+            appConfig.constants.METAMASK_GAS_LIMIT_TRANSFER_FT,
+          );
+
+          //@TODO implement this flow to get readable results to show them to the user
+          // in order to read the contract call results, you will need to query the contract call's results form a mirror node using the transaction id
+          // after getting the contract call results, use ethers and abi.decode to decode the call_result
+
+          console.log("txId", txId);
+          //setTxId(txId as string);
+        }}
+      >
+        Send
+      </Button>
+    </VStack>
+  );
+}
