+++ src/layout/Layout.tsx
@@ -1,22 +1,20 @@
 "use client";
 
-import { Box, Flex } from "@chakra-ui/react";
+import { Box, Divider, Flex } from "@chakra-ui/react";
 import type { ReactNode } from "react";
 import Sidebar from "./sidebar/Sidebar";
 import Topbar from "./topbar/Topbar";
 import Footer from "./footer/Footer";
 
 import { ChakraProvider } from "@chakra-ui/react";
-import theme from "../theme/Theme";
-import { Fonts } from "../components/Fonts";
-import { Focus } from "../components/Focus";
-import { ScrollBar } from "../components/Scrollbar";
+import theme from "@/theme/Theme";
+import { Fonts } from "@/components/Fonts";
+import { Focus } from "@/components/Focus";
+import { ScrollBar } from "@/components/Scrollbar";
 
-import { AllWalletsProvider } from "../services/wallets/AllWalletsProvider";
+import { AllWalletsProvider } from "@/services/wallets/AllWalletsProvider";
 
-import TestTransactions from "../views/TestTransactions";
-import TestFunctionCall from "@/views/TestFunctionCall";
-import TransferFungibleToken from "@/views/TransferFungibleToken";
+import EIP3643 from "@/views/eip3643/EIP3643";
 
 interface LayoutProps {
   children: ReactNode;
@@ -28,16 +26,13 @@ const Layout = ({ children }: LayoutProps) => {
       <Focus />
       <Fonts />
       <AllWalletsProvider>
-        <Box h="100vh" w="100%" overflowX="hidden">
+        <Box>
           <Topbar />
-
-          <Flex transition="transform 0.3s, width 0.3s" h="calc(100vh - 64px)">
+          <Flex>
             <Sidebar />
-            <Box w="100%" py={9} px={10} minH="100%" bgColor="brand.gray5">
-              {/* {children} TODO: Make it right*/}
-              <TestTransactions></TestTransactions>
-              <TestFunctionCall />
-              <TransferFungibleToken />
+            <Box p={10}>
+              {/* {children} TODO: we can't use next.js routing due to build error of 'crypto' module */}
+              <EIP3643 />
             </Box>
           </Flex>
           <Footer />
